+++ examples/egg/app/graphql/module/test/schema.ts
@@ -1,0 +1,17 @@
+export const testTypeDefs = `#graphql
+
+type Test {
+    id: String
+
+    name: String
+}
+
+type TestList {
+    count: Int
+    data: [Test]
+}
+
+type Query {
+    testList: TestList
+}
+`;