+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @graphql-yoga/nestjs
 
+## 3.10.8
+
+### Patch Changes
+
+- Updated dependencies
+  [[`ed344ea`](https://github.com/dotansimha/graphql-yoga/commit/ed344ea7fed85163eba0c636b5b7f64c482ce150)]:
+  - graphql-yoga@5.10.8
+
+## 3.10.7
+
+### Patch Changes
+
+- Updated dependencies
+  [[`8fee214`](https://github.com/dotansimha/graphql-yoga/commit/8fee214eac4be3e9e18935502d32de69f0a1484e),
+  [`1df4912`](https://github.com/dotansimha/graphql-yoga/commit/1df4912693fad5efaf2fe99bf0433c9abc829004)]:
+  - graphql-yoga@5.10.7
+
 ## 3.10.6
 
 ### Patch Changes