+++ packages/plugins/sofa/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @graphql-yoga/plugin-sofa
 
+## 3.10.8
+
+### Patch Changes
+
+- Updated dependencies
+  [[`ed344ea`](https://github.com/dotansimha/graphql-yoga/commit/ed344ea7fed85163eba0c636b5b7f64c482ce150)]:
+  - graphql-yoga@5.10.8
+
+## 3.10.7
+
+### Patch Changes
+
+- [#3559](https://github.com/dotansimha/graphql-yoga/pull/3559)
+  [`5620283`](https://github.com/dotansimha/graphql-yoga/commit/56202838e3f6ce0706ee2890b539bd5339dc0598)
+  Thanks [@ardatan](https://github.com/ardatan)! - dependencies updates:
+
+  - Updated dependency [`sofa-api@^0.18.8` ↗︎](https://www.npmjs.com/package/sofa-api/v/0.18.8)
+    (from `^0.18.7`, in `dependencies`)
+
+- [#3559](https://github.com/dotansimha/graphql-yoga/pull/3559)
+  [`5620283`](https://github.com/dotansimha/graphql-yoga/commit/56202838e3f6ce0706ee2890b539bd5339dc0598)
+  Thanks [@ardatan](https://github.com/ardatan)! - Fix the issue when SOFA returns 404 response from
+  error extensions returned by a resolver, it will cause the server to continue the request handling
+  with Yoga but instead it should return the response with 404 and the body SOFA returns.
+
+- Updated dependencies
+  [[`8fee214`](https://github.com/dotansimha/graphql-yoga/commit/8fee214eac4be3e9e18935502d32de69f0a1484e),
+  [`1df4912`](https://github.com/dotansimha/graphql-yoga/commit/1df4912693fad5efaf2fe99bf0433c9abc829004)]:
+  - graphql-yoga@5.10.7
+
 ## 3.10.6
 
 ### Patch Changes